move allocator shim logic into its own file
